+++ State_machine/src/S1_STATE/S1_STATE.cpp
@@ -62,7 +62,7 @@ boolean debounceButton(boolean state,int buttonPin)
   return stateNow;
 }
 
-boolean compareDates(int final_year, int final_month, int final_day, int final_hour, int final_minute, int current_year, int current_month, int current_day, int current_hour, int current_minute){
+boolean compareDates(int final_year, uint8_t final_month, uint8_t final_day, uint8_t final_hour, uint8_t final_minute, int current_year, uint8_t current_month, uint8_t current_day, uint8_t current_hour, uint8_t current_minute){
 
   if(final_year == current_year){
 
@@ -107,7 +107,7 @@ boolean compareDates(int final_year, int final_month, int final_day, int final_h
 }
 
 // Esta funcion retorna True si falta un mes o menos 
-boolean lessThanMonth(int final_year, int final_month, int current_year, int current_month){
+boolean lessThanMonth(int final_year, uint8_t final_month, int current_year, uint8_t current_month){
 
   if(final_year==current_year){
 
@@ -132,7 +132,7 @@ boolean lessThanMonth(int final_year, int final_month, int current_year, int cur
 
 }
 
-void countdown_date(int final_day, int final_hour, int final_minute, int current_day,int current_hour, int current_minute){
+void countdown_date(uint8_t final_day, uint8_t final_hour, uint8_t final_minute, uint8_t current_day,uint8_t current_hour, uint8_t current_minute){
 int number_days=final_day - current_day;
 int number_hours= final_hour - current_hour;
 int number_minutes= final_minute - current_minute;
